@@ -3,7 +3,7 @@ description: A new Flutter project.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -11,35 +11,37 @@ dependencies:
   intl: ^0.16.0
   url_launcher: 5.4.2
   get_it: 4.0.1
-  auto_route: 0.6.9
+  auto_route: 1.0.2
   cupertino_icons: ^0.1.3
-  connectivity: 0.4.8+2
-  cached_network_image: ^2.2.0
-  json_annotation: 3.1.0
-  shimmer: 1.1.1
-  package_info: ^0.4.0+18
-  fluttertoast: ^4.0.0
-  smooth_page_indicator: 0.1.5
-  freezed: 0.12.2
-  flutter_bloc: 6.0.0
+  connectivity: 3.0.6
+  cached_network_image: 2.5.1
+  json_annotation: 4.1.0
+  shimmer: 2.0.0
+  package_info: 2.0.2
+  fluttertoast: 8.0.8
+  smooth_page_indicator: 1.0.0+2
+  freezed: 0.14.1+2
+  flutter_bloc: 8.0.0
   cloud_firestore: 0.13.7
   firebase_crashlytics: 0.1.3+3
   firebase_auth: 0.16.1
   firebase_core: 0.4.5
-  razorpay_flutter: 1.1.3
-  country_code_picker: ^2.0.1
+  razorpay_flutter: 1.2.7
+  country_code_picker: 2.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  build_runner: 1.10.3
-  json_serializable: ^3.5.0
-  auto_route_generator: 0.6.10
+  build_runner:
+  json_serializable:
+  auto_route_generator:
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
-dependency_overrides:
-  analyzer: 0.40.4
-# The following section is specific to Flutter.
+#dependency_overrides:
+#  analyzer:
+#  meta:
+#  build_resolvers:
+## The following section is specific to Flutter.
 flutter:
   uses-material-design: true
 
