@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'address_card_state.dart';
 
@@ -9,81 +9,79 @@ part of 'address_card_state.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$AddressCardStateTearOff {
   const _$AddressCardStateTearOff();
 
-// ignore: unused_element
   Idle idle() {
     return const Idle();
   }
 
-// ignore: unused_element
   EditLoading editLoading() {
     return const EditLoading();
   }
 
-// ignore: unused_element
   SetDefaultLoading setDefaultLoading() {
     return const SetDefaultLoading();
   }
 
-// ignore: unused_element
   Error error(String message) {
     return Error(
       message,
     );
   }
 
-// ignore: unused_element
   Successful successful() {
     return const Successful();
   }
 }
 
 /// @nodoc
-// ignore: unused_element
 const $AddressCardState = _$AddressCardStateTearOff();
 
 /// @nodoc
 mixin _$AddressCardState {
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result editLoading(),
-    @required Result setDefaultLoading(),
-    @required Result error(String message),
-    @required Result successful(),
-  });
-
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() editLoading,
+    required TResult Function() setDefaultLoading,
+    required TResult Function(String message) error,
+    required TResult Function() successful,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result editLoading(),
-    Result setDefaultLoading(),
-    Result error(String message),
-    Result successful(),
-    @required Result orElse(),
-  });
-
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? editLoading,
+    TResult Function()? setDefaultLoading,
+    TResult Function(String message)? error,
+    TResult Function()? successful,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result editLoading(EditLoading value),
-    @required Result setDefaultLoading(SetDefaultLoading value),
-    @required Result error(Error value),
-    @required Result successful(Successful value),
-  });
-
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(EditLoading value) editLoading,
+    required TResult Function(SetDefaultLoading value) setDefaultLoading,
+    required TResult Function(Error value) error,
+    required TResult Function(Successful value) successful,
+  }) =>
+      throw _privateConstructorUsedError;
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result editLoading(EditLoading value),
-    Result setDefaultLoading(SetDefaultLoading value),
-    Result error(Error value),
-    Result successful(Successful value),
-    @required Result orElse(),
-  });
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(EditLoading value)? editLoading,
+    TResult Function(SetDefaultLoading value)? setDefaultLoading,
+    TResult Function(Error value)? error,
+    TResult Function(Successful value)? successful,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -99,7 +97,6 @@ class _$AddressCardStateCopyWithImpl<$Res>
   _$AddressCardStateCopyWithImpl(this._value, this._then);
 
   final AddressCardState _value;
-
   // ignore: unused_field
   final $Res Function(AddressCardState) _then;
 }
@@ -139,32 +136,26 @@ class _$Idle implements Idle {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result editLoading(),
-    @required Result setDefaultLoading(),
-    @required Result error(String message),
-    @required Result successful(),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() editLoading,
+    required TResult Function() setDefaultLoading,
+    required TResult Function(String message) error,
+    required TResult Function() successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return idle();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result editLoading(),
-    Result setDefaultLoading(),
-    Result error(String message),
-    Result successful(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? editLoading,
+    TResult Function()? setDefaultLoading,
+    TResult Function(String message)? error,
+    TResult Function()? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (idle != null) {
       return idle();
     }
@@ -173,32 +164,26 @@ class _$Idle implements Idle {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result editLoading(EditLoading value),
-    @required Result setDefaultLoading(SetDefaultLoading value),
-    @required Result error(Error value),
-    @required Result successful(Successful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(EditLoading value) editLoading,
+    required TResult Function(SetDefaultLoading value) setDefaultLoading,
+    required TResult Function(Error value) error,
+    required TResult Function(Successful value) successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return idle(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result editLoading(EditLoading value),
-    Result setDefaultLoading(SetDefaultLoading value),
-    Result error(Error value),
-    Result successful(Successful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(EditLoading value)? editLoading,
+    TResult Function(SetDefaultLoading value)? setDefaultLoading,
+    TResult Function(Error value)? error,
+    TResult Function(Successful value)? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (idle != null) {
       return idle(this);
     }
@@ -248,32 +233,26 @@ class _$EditLoading implements EditLoading {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result editLoading(),
-    @required Result setDefaultLoading(),
-    @required Result error(String message),
-    @required Result successful(),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() editLoading,
+    required TResult Function() setDefaultLoading,
+    required TResult Function(String message) error,
+    required TResult Function() successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return editLoading();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result editLoading(),
-    Result setDefaultLoading(),
-    Result error(String message),
-    Result successful(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? editLoading,
+    TResult Function()? setDefaultLoading,
+    TResult Function(String message)? error,
+    TResult Function()? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (editLoading != null) {
       return editLoading();
     }
@@ -282,32 +261,26 @@ class _$EditLoading implements EditLoading {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result editLoading(EditLoading value),
-    @required Result setDefaultLoading(SetDefaultLoading value),
-    @required Result error(Error value),
-    @required Result successful(Successful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(EditLoading value) editLoading,
+    required TResult Function(SetDefaultLoading value) setDefaultLoading,
+    required TResult Function(Error value) error,
+    required TResult Function(Successful value) successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return editLoading(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result editLoading(EditLoading value),
-    Result setDefaultLoading(SetDefaultLoading value),
-    Result error(Error value),
-    Result successful(Successful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(EditLoading value)? editLoading,
+    TResult Function(SetDefaultLoading value)? setDefaultLoading,
+    TResult Function(Error value)? error,
+    TResult Function(Successful value)? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (editLoading != null) {
       return editLoading(this);
     }
@@ -357,32 +330,26 @@ class _$SetDefaultLoading implements SetDefaultLoading {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result editLoading(),
-    @required Result setDefaultLoading(),
-    @required Result error(String message),
-    @required Result successful(),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() editLoading,
+    required TResult Function() setDefaultLoading,
+    required TResult Function(String message) error,
+    required TResult Function() successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return setDefaultLoading();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result editLoading(),
-    Result setDefaultLoading(),
-    Result error(String message),
-    Result successful(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? editLoading,
+    TResult Function()? setDefaultLoading,
+    TResult Function(String message)? error,
+    TResult Function()? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (setDefaultLoading != null) {
       return setDefaultLoading();
     }
@@ -391,32 +358,26 @@ class _$SetDefaultLoading implements SetDefaultLoading {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result editLoading(EditLoading value),
-    @required Result setDefaultLoading(SetDefaultLoading value),
-    @required Result error(Error value),
-    @required Result successful(Successful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(EditLoading value) editLoading,
+    required TResult Function(SetDefaultLoading value) setDefaultLoading,
+    required TResult Function(Error value) error,
+    required TResult Function(Successful value) successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return setDefaultLoading(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result editLoading(EditLoading value),
-    Result setDefaultLoading(SetDefaultLoading value),
-    Result error(Error value),
-    Result successful(Successful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(EditLoading value)? editLoading,
+    TResult Function(SetDefaultLoading value)? setDefaultLoading,
+    TResult Function(Error value)? error,
+    TResult Function(Successful value)? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (setDefaultLoading != null) {
       return setDefaultLoading(this);
     }
@@ -432,7 +393,6 @@ abstract class SetDefaultLoading implements AddressCardState {
 abstract class $ErrorCopyWith<$Res> {
   factory $ErrorCopyWith(Error value, $Res Function(Error) then) =
       _$ErrorCopyWithImpl<$Res>;
-
   $Res call({String message});
 }
 
@@ -447,17 +407,20 @@ class _$ErrorCopyWithImpl<$Res> extends _$AddressCardStateCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object message = freezed,
+    Object? message = freezed,
   }) {
     return _then(Error(
-      message == freezed ? _value.message : message as String,
+      message == freezed
+          ? _value.message
+          : message // ignore: cast_nullable_to_non_nullable
+              as String,
     ));
   }
 }
 
 /// @nodoc
 class _$Error implements Error {
-  const _$Error(this.message) : assert(message != null);
+  const _$Error(this.message);
 
   @override
   final String message;
@@ -479,38 +442,33 @@ class _$Error implements Error {
   int get hashCode =>
       runtimeType.hashCode ^ const DeepCollectionEquality().hash(message);
 
+  @JsonKey(ignore: true)
   @override
   $ErrorCopyWith<Error> get copyWith =>
       _$ErrorCopyWithImpl<Error>(this, _$identity);
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result editLoading(),
-    @required Result setDefaultLoading(),
-    @required Result error(String message),
-    @required Result successful(),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() editLoading,
+    required TResult Function() setDefaultLoading,
+    required TResult Function(String message) error,
+    required TResult Function() successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return error(message);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result editLoading(),
-    Result setDefaultLoading(),
-    Result error(String message),
-    Result successful(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? editLoading,
+    TResult Function()? setDefaultLoading,
+    TResult Function(String message)? error,
+    TResult Function()? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (error != null) {
       return error(message);
     }
@@ -519,32 +477,26 @@ class _$Error implements Error {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result editLoading(EditLoading value),
-    @required Result setDefaultLoading(SetDefaultLoading value),
-    @required Result error(Error value),
-    @required Result successful(Successful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(EditLoading value) editLoading,
+    required TResult Function(SetDefaultLoading value) setDefaultLoading,
+    required TResult Function(Error value) error,
+    required TResult Function(Successful value) successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return error(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result editLoading(EditLoading value),
-    Result setDefaultLoading(SetDefaultLoading value),
-    Result error(Error value),
-    Result successful(Successful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(EditLoading value)? editLoading,
+    TResult Function(SetDefaultLoading value)? setDefaultLoading,
+    TResult Function(Error value)? error,
+    TResult Function(Successful value)? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (error != null) {
       return error(this);
     }
@@ -555,9 +507,9 @@ class _$Error implements Error {
 abstract class Error implements AddressCardState {
   const factory Error(String message) = _$Error;
 
-  String get message;
-
-  $ErrorCopyWith<Error> get copyWith;
+  String get message => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $ErrorCopyWith<Error> get copyWith => throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -597,32 +549,26 @@ class _$Successful implements Successful {
 
   @override
   @optionalTypeArgs
-  Result when<Result extends Object>({
-    @required Result idle(),
-    @required Result editLoading(),
-    @required Result setDefaultLoading(),
-    @required Result error(String message),
-    @required Result successful(),
+  TResult when<TResult extends Object?>({
+    required TResult Function() idle,
+    required TResult Function() editLoading,
+    required TResult Function() setDefaultLoading,
+    required TResult Function(String message) error,
+    required TResult Function() successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return successful();
   }
 
   @override
   @optionalTypeArgs
-  Result maybeWhen<Result extends Object>({
-    Result idle(),
-    Result editLoading(),
-    Result setDefaultLoading(),
-    Result error(String message),
-    Result successful(),
-    @required Result orElse(),
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function()? idle,
+    TResult Function()? editLoading,
+    TResult Function()? setDefaultLoading,
+    TResult Function(String message)? error,
+    TResult Function()? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (successful != null) {
       return successful();
     }
@@ -631,32 +577,26 @@ class _$Successful implements Successful {
 
   @override
   @optionalTypeArgs
-  Result map<Result extends Object>({
-    @required Result idle(Idle value),
-    @required Result editLoading(EditLoading value),
-    @required Result setDefaultLoading(SetDefaultLoading value),
-    @required Result error(Error value),
-    @required Result successful(Successful value),
+  TResult map<TResult extends Object?>({
+    required TResult Function(Idle value) idle,
+    required TResult Function(EditLoading value) editLoading,
+    required TResult Function(SetDefaultLoading value) setDefaultLoading,
+    required TResult Function(Error value) error,
+    required TResult Function(Successful value) successful,
   }) {
-    assert(idle != null);
-    assert(editLoading != null);
-    assert(setDefaultLoading != null);
-    assert(error != null);
-    assert(successful != null);
     return successful(this);
   }
 
   @override
   @optionalTypeArgs
-  Result maybeMap<Result extends Object>({
-    Result idle(Idle value),
-    Result editLoading(EditLoading value),
-    Result setDefaultLoading(SetDefaultLoading value),
-    Result error(Error value),
-    Result successful(Successful value),
-    @required Result orElse(),
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(Idle value)? idle,
+    TResult Function(EditLoading value)? editLoading,
+    TResult Function(SetDefaultLoading value)? setDefaultLoading,
+    TResult Function(Error value)? error,
+    TResult Function(Successful value)? successful,
+    required TResult orElse(),
   }) {
-    assert(orElse != null);
     if (successful != null) {
       return successful(this);
     }
