@@ -1,12 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:fluttercommerce/features/init/screen/splash_screen.dart';
-import 'package:fluttercommerce/features/order/bloc/account_details_cubit.dart';
-import 'package:fluttercommerce/routes/router.gr.dart';
-import 'package:fluttercommerce/res/app_theme.dart';
 import 'package:fluttercommerce/di/di.dart';
+import 'package:fluttercommerce/features/order/bloc/account_details_cubit.dart';
+import 'package:fluttercommerce/res/app_theme.dart';
+import 'package:fluttercommerce/routes/router.gr.dart';
 
 class App extends StatefulWidget {
-  const App({Key key}) : super(key: key);
+  const App({Key? key}) : super(key: key);
 
   @override
   _AppState createState() => _AppState();
@@ -22,12 +21,12 @@ class _AppState extends State<App> {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    final _appRouter = AppRouter();
+    return MaterialApp.router(
       theme: AppTheme.appTheme(),
       debugShowCheckedModeBanner: false,
-      onGenerateRoute: AppRouter(),
-      initialRoute: Routes.splashScreen,
-      home: SplashScreen(),
+      routerDelegate: _appRouter.delegate(),
+      routeInformationParser: _appRouter.defaultRouteParser(),
     );
   }
 }
