@@ -1,18 +1,15 @@
+import 'package:json_annotation/json_annotation.dart';
+
+part 'faq_model.g.dart';
+
+@JsonSerializable()
 class FaqModel {
-  String question;
-  String answer;
+  String? question;
+  String? answer;
 
   FaqModel({this.question, this.answer});
 
-  factory FaqModel.fromDocument(json) {
-    return FaqModel(
-      question: json['question'] as String,
-      answer: json['answer'] as String,
-    );
-  }
+  factory FaqModel.fromDocument(json) => _$FaqModelFromJson(json);
 
-  Map<String, dynamic> toMap() => <String, dynamic>{
-        'question': question,
-        'answer': answer,
-      };
+  Map<String, dynamic> toMap() => _$FaqModelToJson(this);
 }
