@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'account_details_state.dart';
 
@@ -9,13 +9,15 @@ part of 'account_details_state.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$AccountDetailsStateTearOff {
   const _$AccountDetailsStateTearOff();
 
-// ignore: unused_element
   _AccountDetailsState call(
-      {AccountDetails accountDetails, Address selectedAddress}) {
+      {AccountDetails? accountDetails, Address? selectedAddress}) {
     return _AccountDetailsState(
       accountDetails: accountDetails,
       selectedAddress: selectedAddress,
@@ -24,16 +26,16 @@ class _$AccountDetailsStateTearOff {
 }
 
 /// @nodoc
-// ignore: unused_element
 const $AccountDetailsState = _$AccountDetailsStateTearOff();
 
 /// @nodoc
 mixin _$AccountDetailsState {
-  AccountDetails get accountDetails;
+  AccountDetails? get accountDetails => throw _privateConstructorUsedError;
+  Address? get selectedAddress => throw _privateConstructorUsedError;
 
-  Address get selectedAddress;
-
-  $AccountDetailsStateCopyWith<AccountDetailsState> get copyWith;
+  @JsonKey(ignore: true)
+  $AccountDetailsStateCopyWith<AccountDetailsState> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -41,8 +43,7 @@ abstract class $AccountDetailsStateCopyWith<$Res> {
   factory $AccountDetailsStateCopyWith(
           AccountDetailsState value, $Res Function(AccountDetailsState) then) =
       _$AccountDetailsStateCopyWithImpl<$Res>;
-
-  $Res call({AccountDetails accountDetails, Address selectedAddress});
+  $Res call({AccountDetails? accountDetails, Address? selectedAddress});
 }
 
 /// @nodoc
@@ -51,22 +52,23 @@ class _$AccountDetailsStateCopyWithImpl<$Res>
   _$AccountDetailsStateCopyWithImpl(this._value, this._then);
 
   final AccountDetailsState _value;
-
   // ignore: unused_field
   final $Res Function(AccountDetailsState) _then;
 
   @override
   $Res call({
-    Object accountDetails = freezed,
-    Object selectedAddress = freezed,
+    Object? accountDetails = freezed,
+    Object? selectedAddress = freezed,
   }) {
     return _then(_value.copyWith(
       accountDetails: accountDetails == freezed
           ? _value.accountDetails
-          : accountDetails as AccountDetails,
+          : accountDetails // ignore: cast_nullable_to_non_nullable
+              as AccountDetails?,
       selectedAddress: selectedAddress == freezed
           ? _value.selectedAddress
-          : selectedAddress as Address,
+          : selectedAddress // ignore: cast_nullable_to_non_nullable
+              as Address?,
     ));
   }
 }
@@ -77,9 +79,8 @@ abstract class _$AccountDetailsStateCopyWith<$Res>
   factory _$AccountDetailsStateCopyWith(_AccountDetailsState value,
           $Res Function(_AccountDetailsState) then) =
       __$AccountDetailsStateCopyWithImpl<$Res>;
-
   @override
-  $Res call({AccountDetails accountDetails, Address selectedAddress});
+  $Res call({AccountDetails? accountDetails, Address? selectedAddress});
 }
 
 /// @nodoc
@@ -95,16 +96,18 @@ class __$AccountDetailsStateCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object accountDetails = freezed,
-    Object selectedAddress = freezed,
+    Object? accountDetails = freezed,
+    Object? selectedAddress = freezed,
   }) {
     return _then(_AccountDetailsState(
       accountDetails: accountDetails == freezed
           ? _value.accountDetails
-          : accountDetails as AccountDetails,
+          : accountDetails // ignore: cast_nullable_to_non_nullable
+              as AccountDetails?,
       selectedAddress: selectedAddress == freezed
           ? _value.selectedAddress
-          : selectedAddress as Address,
+          : selectedAddress // ignore: cast_nullable_to_non_nullable
+              as Address?,
     ));
   }
 }
@@ -114,9 +117,9 @@ class _$_AccountDetailsState implements _AccountDetailsState {
   _$_AccountDetailsState({this.accountDetails, this.selectedAddress});
 
   @override
-  final AccountDetails accountDetails;
+  final AccountDetails? accountDetails;
   @override
-  final Address selectedAddress;
+  final Address? selectedAddress;
 
   @override
   String toString() {
@@ -141,6 +144,7 @@ class _$_AccountDetailsState implements _AccountDetailsState {
       const DeepCollectionEquality().hash(accountDetails) ^
       const DeepCollectionEquality().hash(selectedAddress);
 
+  @JsonKey(ignore: true)
   @override
   _$AccountDetailsStateCopyWith<_AccountDetailsState> get copyWith =>
       __$AccountDetailsStateCopyWithImpl<_AccountDetailsState>(
@@ -149,15 +153,15 @@ class _$_AccountDetailsState implements _AccountDetailsState {
 
 abstract class _AccountDetailsState implements AccountDetailsState {
   factory _AccountDetailsState(
-      {AccountDetails accountDetails,
-      Address selectedAddress}) = _$_AccountDetailsState;
+      {AccountDetails? accountDetails,
+      Address? selectedAddress}) = _$_AccountDetailsState;
 
   @override
-  AccountDetails get accountDetails;
-
+  AccountDetails? get accountDetails => throw _privateConstructorUsedError;
   @override
-  Address get selectedAddress;
-
+  Address? get selectedAddress => throw _privateConstructorUsedError;
   @override
-  _$AccountDetailsStateCopyWith<_AccountDetailsState> get copyWith;
+  @JsonKey(ignore: true)
+  _$AccountDetailsStateCopyWith<_AccountDetailsState> get copyWith =>
+      throw _privateConstructorUsedError;
 }
