@@ -1,7 +1,10 @@
 class Validator {
-  String validateName(String value) {
-    var pattern = r'(^[a-zA-Z ]*$)';
-    var regExp = new RegExp(pattern);
+  String? validateName(String? value) {
+    const pattern = r'(^[a-zA-Z ]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "Name is Required";
     } else if (!regExp.hasMatch(value.trim())) {
@@ -10,9 +13,12 @@ class Validator {
     return null;
   }
 
-  String validateMobile(String value) {
-    var pattern = r'(^[0-9]*$)';
-    var regExp = new RegExp(pattern);
+  String? validateMobile(String? value) {
+    const pattern = r'(^[0-9]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "Phone number is Required";
     } else if (value.length != 10) {
@@ -23,9 +29,12 @@ class Validator {
     return null;
   }
 
-  String validate6DigitCode(String value) {
-    var pattern = r'(^[0-9]*$)';
-    var regExp = new RegExp(pattern);
+  String? validate6DigitCode(String? value) {
+    const pattern = r'(^[0-9]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "Passcode is Required";
     } else if (value.length != 6) {
@@ -36,9 +45,12 @@ class Validator {
     return null;
   }
 
-  String validate6DigitToken(String value) {
-    var pattern = r'(^[0-9]*$)';
-    var regExp = new RegExp(pattern);
+  String? validate6DigitToken(String? value) {
+    const pattern = r'(^[0-9]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "Please enter a valid token";
     } else if (value.length != 6) {
@@ -49,9 +61,12 @@ class Validator {
     return null;
   }
 
-  String validatePassCode(String value) {
-    var pattern = r'(^[0-9]*$)';
-    var regExp = new RegExp(pattern);
+  String? validatePassCode(String? value) {
+    const pattern = r'(^[0-9]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "Passcode is Required";
     } else if (value.length != 6) {
@@ -62,9 +77,12 @@ class Validator {
     return null;
   }
 
-  String passwordDoNotMatch(String value, String confirmValue) {
-    var pattern = r'(^[0-9]*$)';
-    var regExp = new RegExp(pattern);
+  String? passwordDoNotMatch(String? value, String? confirmValue) {
+    const pattern = r'(^[0-9]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "PassCode is Required";
     } else if (value.length != 6) {
@@ -77,10 +95,13 @@ class Validator {
     return null;
   }
 
-  String validateEmail(String value) {
-    var pattern =
+  String? validateEmail(String? value) {
+    const pattern =
         r'^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$';
-    var regExp = new RegExp(pattern);
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "Email is Required";
     } else if (!regExp.hasMatch(value)) {
@@ -90,9 +111,12 @@ class Validator {
     }
   }
 
-  String validatePinCode(String value) {
-    var pattern = r'(^[0-9]*$)';
-    var regExp = new RegExp(pattern);
+  String? validatePinCode(String? value) {
+    const pattern = r'(^[0-9]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "Pincode is Required";
     } else if (value.length != 6) {
@@ -103,9 +127,12 @@ class Validator {
     return null;
   }
 
-  String validateText(String value, String text) {
-    var pattern = r'(^[a-zA-Z ]*$)';
-    var regExp = new RegExp(pattern);
+  String? validateText(String? value, String? text) {
+    const pattern = r'(^[a-zA-Z ]*$)';
+    final regExp = RegExp(pattern);
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "$text is Required";
     } else if (!regExp.hasMatch(value.trim())) {
@@ -114,7 +141,10 @@ class Validator {
     return null;
   }
 
-  String validateNumber(String value, String desc) {
+  String? validateNumber(String? value, String? desc) {
+    if (value == null) {
+      return null;
+    }
     if (value.isEmpty) {
       return "$desc is Required";
     }
