@@ -6,14 +6,12 @@ import 'package:fluttercommerce/features/common/models/account_details_model.dar
 import 'package:fluttercommerce/core/utils/validator.dart';
 
 class AddAccountDetailsCubit extends Cubit<AddAccountDetailsState> {
-  FirestoreRepository _firebaseRepo;
-
-  AuthRepository _authRepo;
-
   AddAccountDetailsCubit(this._firebaseRepo, this._authRepo)
       : super(const AddAccountDetailsState.idle());
+
+  final FirestoreRepository _firebaseRepo;
+  final AuthRepository _authRepo;
   final Validator _validator = Validator();
-  AccountDetails _accountDetails;
 
   void validateButton(String name) {
     if (_validator.validateName(name) == null) {
@@ -25,17 +23,15 @@ class AddAccountDetailsCubit extends Cubit<AddAccountDetailsState> {
 
   Future<void> loadPreviousData() async {
     emit(const AddAccountDetailsState.loading());
-    _accountDetails = await _firebaseRepo.fetchUserDetails();
+    AccountDetails _accountDetails = await _firebaseRepo.fetchUserDetails();
     emit(AddAccountDetailsState.editData(_accountDetails));
     validateButton(_accountDetails.name);
   }
 
-  saveData(String name, {bool isEdit = false}) async {
-    if (isEdit) {
-      _accountDetails = AccountDetails();
-    }
-    _accountDetails.name = name;
-    _accountDetails.phoneNumber = await _authRepo.getPhoneNumber();
+  Future<void> saveData(String name, {bool isEdit = false}) async {
+    final AccountDetails _accountDetails = AccountDetails(
+        name: name, phoneNumber: await _authRepo.getPhoneNumber());
+
     emit(const AddAccountDetailsState.saveDataLoading());
     await _firebaseRepo.addUserDetails(_accountDetails);
     await _authRepo.setAccountDetails(displayName: _accountDetails.name);
