@@ -1,11 +1,10 @@
 import 'package:flutter/material.dart';
 
 class WidgetVisibility extends StatelessWidget {
+  WidgetVisibility({required this.visible, required this.child});
+
   final bool visible;
   final Widget child;
-
-  WidgetVisibility({@required this.visible, @required this.child})
-      : assert(visible != null && child != null);
 
   @override
   Widget build(BuildContext context) {
