@@ -6,6 +6,6 @@ part 'account_details_state.freezed.dart';
 @freezed
 abstract class AccountDetailsState with _$AccountDetailsState {
   factory AccountDetailsState(
-      {AccountDetails accountDetails,
-      Address selectedAddress}) = _AccountDetailsState;
+      {AccountDetails? accountDetails,
+      Address? selectedAddress}) = _AccountDetailsState;
 }
