@@ -1,4 +1,3 @@
-import 'package:auto_route/auto_route_annotations.dart';
 import 'package:fluttercommerce/features/address/screen/add_address_screen.dart';
 import 'package:fluttercommerce/features/address/screen/my_address_screen.dart';
 import 'package:fluttercommerce/features/checkout/screen/cart_screen.dart';
@@ -12,7 +11,7 @@ import 'package:fluttercommerce/features/product/screen/all_product_list_screen.
 import 'package:fluttercommerce/features/product/screen/product_detail_page.dart';
 import 'package:fluttercommerce/features/search/screen/SearchScreen.dart';
 import 'package:fluttercommerce/features/user/screen/add_user_detail_screen.dart';
-
+import 'package:auto_route/annotations.dart';
 /*
 * Run this command in terminal
 * flutter packages pub run build_runner build
@@ -41,6 +40,6 @@ import 'package:fluttercommerce/features/user/screen/add_user_detail_screen.dart
     CupertinoRoute(page: AddAddressScreen),
     CupertinoRoute(page: MyOrdersScreen),
   ],
-  routesClassName: "Routes",
+  replaceInRouteName: "Routes",
 )
 class $AppRouter {}
